fix: Make policies optional
Sometimes a policy reasoner does not wish to use the `policy reasoner`
to determine its policy. We could use empty policies for those cases,
but it is probably nicer, to allow policies to be nullable, and let the
reasoner handle that case.

Most policy reasoner connectors will end up doing a rejection anyway,
but now they can choose not to if that is better for them.