Started major refactor into a library
This branch will contain the work on the `policy-reasoner` to turn the
repository into a library instead of a binary. This is done to remove
the Brane-specific parts of it, and put that into the Brane repo. As
such, end binaries will be in the codebase of the respective project;
the exception is the existing binaries (kind of), which are now
examples.

So far, moved the most important parts. Some work still needs to be done
to re-introduce the appropriate logging, and fixing the POSIX reasoner.
The rest -i.e., Brane parts and the store-part- need to be moved to
independent repositories. The same applies to the eFLINT base policy (or
that has to live in Brane land, idk).